Revise build flow

+++ src/rendering/Panel.ts
@@ -92,7 +92,7 @@ export class Panel {
 
             for (var i = 0; i < 3; i++) {
                 const geometry = new THREE.BufferGeometry();
-                geometry.addAttribute('position', new THREE.BufferAttribute(axisGeometry.getAttribute('position').array.subarray(i * 6 + 3, i * 6 + 6), 3));
+                geometry.addAttribute('position', new THREE.BufferAttribute((axisGeometry.getAttribute('position').array as Float32Array).subarray(i * 6 + 3, i * 6 + 6), 3));
                 this.axisObject.add(new THREE.Points(geometry, new THREE.PointsMaterial({
                     alphaTest: 0.17        // vaccarium.TODO: this is a horrible hack, see SOverflow #27042683
                 })));
@@ -152,7 +152,7 @@ export class Panel {
 
             const geometry = new THREE.BufferGeometry();
             geometry.addAttribute('position', new THREE.BufferAttribute(Pool.get(Float32Array, 3), 3));
-            geometry.getAttribute('position').array[pos] = distance;
+            (geometry.getAttribute('position').array[pos] as Float32Array) = distance;
 
             const textObject = new THREE.Points(geometry, new THREE.PointsMaterial({
                 alphaTest: 0.17        // vaccarium.TODO: this is a horrible hack, see SOverflow #27042683
